requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
django-cors-headers>=3.10.0,<4.0.0
mongoengine>=0.20.0
pymongo>=3.12.0
python-dateutil>=2.8.0
pandas>=1.3.0
PyJWT>=2.4.0 